requirements.txt
Django==5.0.2
django-environ==0.11.2
django-crispy-forms==2.1
crispy-tailwind==0.5.0
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
dj-database-url==2.1.0
django-allauth==0.57.0
python-dateutil==2.8.2